Moved redundant test-prep code into testUtils.ts.
