+++ client/src/main/scripts/components/EditorVideoContainer.spec.tsx
@@ -7,6 +7,7 @@ import {
   EditorVideoContainerState
 } from "./EditorVideoContainer";
 import { Foundation } from "../java2ts/Foundation";
+import { videoFact } from "../utils/testUtils";
 
 const mockFn = jest.fn();
 
@@ -21,65 +22,6 @@ const containerProps: EditorVideoContainerProps = {
     kind: "video",
     range: [2, 4],
     videoId: "U8MV5KDDaxumxZOCJOzExAUAAkSoYNhycVXq7jZ59_0="
-  }
-};
-
-const videoFact: Foundation.VideoFactContent = {
-  youtubeId: "ApTLB76Nmdg",
-  speakers: [
-    {
-      firstname: "Martha",
-      middlename: null,
-      lastname: "Raddatz"
-    }
-  ],
-  transcript: [
-    {
-      idx: 0,
-      word: "Good ",
-      timestamp: 0.75
-    },
-    {
-      idx: 1,
-      word: "evening ",
-      timestamp: 0.96
-    },
-    {
-      idx: 2,
-      word: "I'm ",
-      timestamp: 1.56
-    },
-    {
-      idx: 3,
-      word: "Martha ",
-      timestamp: 1.92
-    },
-    {
-      idx: 4,
-      word: "Raddatz ",
-      timestamp: 2.129
-    },
-    {
-      idx: 5,
-      word: "from ",
-      timestamp: 2.52
-    },
-    {
-      idx: 6,
-      word: "ABC ",
-      timestamp: 3.06
-    },
-    {
-      idx: 7,
-      word: "News. ",
-      timestamp: 3.09
-    }
-  ],
-  fact: {
-    title: "Donald Trump - Hillary Clinton (2/3)",
-    primaryDate: "2016-10-09",
-    primaryDateKind: "recorded",
-    kind: "video"
   }
 };
