+++ client/src/main/scripts/components/TimelineView.spec.tsx
@@ -6,6 +6,7 @@ import {
   TimelineViewState,
   EventHandlers
 } from "./TimelineView";
+import { documentFactLink, timelineItems, videoFactLink } from "../utils/testUtils";
 import { SetFactHandlers } from "./TimelinePreview";
 import { Foundation } from "../java2ts/Foundation";
 
@@ -42,57 +43,6 @@ const initialState: TimelineViewState = {
   hashValues: null,
   hashIsValid: false
 };
-
-const documentFactLink: Foundation.FactLink = {
-  fact: {
-    title: "Amendment 13",
-    primaryDate: "1865-12-06",
-    primaryDateKind: "ratified",
-    kind: "document"
-  },
-  hash: "o_dRqrNJ62wzlgLilTrLxkHqGmvAS9qTpa4z4pjyFqA="
-};
-
-const videoFactLink: Foundation.FactLink = {
-  fact: {
-    title: "Donald Trump - Hillary Clinton (2/3)",
-    primaryDate: "2016-10-09",
-    primaryDateKind: "recorded",
-    kind: "video"
-  },
-  hash: "U8MV5KDDaxumxZOCJOzExAUAAkSoYNhycVXq7jZ59_0="
-};
-
-const timelineItems = [
-  {
-    id: "c7qu-ZE5SuipqSrOO30R3mnAA7K7nJ4fQ4zVIX0A2yg=",
-    idx: "c7qu-ZE5SuipqSrOO30R3mnAA7K7nJ4fQ4zVIX0A2yg=",
-    start: new Date("1788-06-21T00:00:00.000Z"),
-    content: "United States Constitution",
-    kind: "document"
-  },
-  {
-    id: "pMHhbW_I-wquOfoyPFAVQu8DMLMpYVxhGT8R1x71hYA=",
-    idx: "pMHhbW_I-wquOfoyPFAVQu8DMLMpYVxhGT8R1x71hYA=",
-    start: new Date("1791-12-15T00:00:00.000Z"),
-    content: "Bill of Rights",
-    kind: "document"
-  },
-  {
-    id: "-7DeOJAVJUsifUcIaZo7c41pol_guMxR6IEgYv28bHM=",
-    idx: "-7DeOJAVJUsifUcIaZo7c41pol_guMxR6IEgYv28bHM=",
-    start: new Date("1960-09-26T00:00:00.000Z"),
-    content: "John F. Kennedy - Nixon (1/4)",
-    kind: "video"
-  },
-  {
-    id: "bl03RovlxbTZK0yu25_VikP0Y2xSj-J9oFyGUTWIOZQ=",
-    idx: "bl03RovlxbTZK0yu25_VikP0Y2xSj-J9oFyGUTWIOZQ=",
-    start: new Date("1960-10-07T00:00:00.000Z"),
-    content: "John F. Kennedy - Nixon (2/4)",
-    kind: "video"
-  }
-];
 
 test("View loading", () => {
   const containerState: TimelineViewState = {
